Refactor error and message response classes to extend a new base Response class for improved code reuse
