+++ server/src/main/java/com/kalyvianakis/estiator/io/controller/TableController.java
@@ -45,7 +45,7 @@ public class TableController {
         }
 
         tableService.delete(id);
-        MessageResponse response = new MessageResponse("Table deleted for ID: " + id);
+        MessageResponse response = new MessageResponse("Table deleted for ID: " + id, "");
         return ResponseEntity.ok().body(response);
     }
