+++ server/src/main/java/com/kalyvianakis/estiator/io/controller/UserController.java
@@ -141,7 +141,7 @@ public class UserController {
         }
 
         userService.delete(id);
-        MessageResponse response = new MessageResponse("Resource deleted for ID: " + id);
+        MessageResponse response = new MessageResponse("Resource deleted for ID: " + id, "");
         return ResponseEntity.ok().body(response);
     }
