+++ server/src/main/java/com/kalyvianakis/estiator/io/model/ErrorResponse.java
@@ -1,29 +1,7 @@
 package com.kalyvianakis.estiator.io.model;
 
-import java.time.LocalDateTime;
-
-public class ErrorResponse {
-    private String message;
-    private String details;
-
+public class ErrorResponse extends Response {
     public ErrorResponse(String message, String details) {
-        this.message = message;
-        this.details = details;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-
-    public String getDetails() {
-        return details;
-    }
-
-    public void setDetails(String details) {
-        this.details = details;
+        super(message, details);
     }
 }
