+++ server/src/main/java/com/kalyvianakis/estiator/io/model/Response.java
@@ -1,0 +1,27 @@
+package com.kalyvianakis.estiator.io.model;
+
+public class Response {
+    private String message;
+    private String details;
+
+    public Response(String message, String details) {
+        this.message = message;
+        this.details = details;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+
+    public String getDetails() {
+        return details;
+    }
+
+    public void setDetails(String details) {
+        this.details = details;
+    }
+}
